grid tiles updates
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk/source@442 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ java/org.wcs.smart/src/org/wcs/smart/util/ReprojectUtils.java
@@ -61,8 +61,8 @@ public class ReprojectUtils {
 	public static String computeTileId(double x, double y, String destCrsWkt,
 			double originX, double originY, double gridSize) throws Exception{
 		Coordinate c = reproject(x, y, destCrsWkt);
-		int tileidx = (int)Math.floor( (c.x - originX) / gridSize) + 1;
-		int tileidy = (int)Math.floor( (c.y - originY) / gridSize) + 1;
+		long tileidx = (long)Math.floor( (c.x - originX) / gridSize) + 1;
+		long tileidy = (long)Math.floor( (c.y - originY) / gridSize) + 1;
 
 		return tileidx + "_" + tileidy;
 	}
